pyproject.toml
[project]
name = "gemini-batch"
version = "0.1.0"
description = "Command-line tools for managing Google Gemini batch jobs"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "CC0-1.0"}
repository = "https://github.com/7shi/gemini-batch"
dependencies = [
    "google-genai>=1.25.0",
    "rich>=14.0.0",
]

[project.scripts]
gembatch = "gembatch.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gembatch"]
